+++ src/screens/auth/styles.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components/native';
 import { StyleSheet } from 'react-native';
-import { Colors } from '../../../../assets/Colors';
+import { Colors } from '../../../assets/Colors';
 
 export const LoginContainer = styled.View`
   width: 100%;
